Finished implementation of VFH, yay!
Everything is in place to start using it. There are some concerns about lack of (unit) tests, but as it's a relatively simple algorithm, without fancy data structures, there shouldn't be any big problem.

Anyway, I'll improve the documentation and test coverage (which is 0% at the moment) when I finish my real-time systems project.